requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.3.5
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
flask==2.3.3 